0899-RNC Enhancement-&-0954 - HVAC Dealer Portal - Calendar issue in IE
Need to be able to allow dealer to include/exclude residential new construction from their ADVANTAGE offer preferences.  This impacts Dealer Enrollment, Create Offer, Update Offer, and View Offer.

We will be adding a new flag (true / false) of whether to include RNC in the offers to the above flows.

***********************************************************************
0954 - HVAC Dealer Portal - Calendar issue in IE

There is an issue in IE where the calendar control is setting the earliest end date for inclusion offers to 30 days from current date instead of 30 days from the offer begin date. This works fine in Firefox & Chrome.